Move /font folder to /framework/font
