+++ src/framework/components/element/component.js
@@ -20,8 +20,8 @@ import { ImageElement } from './image-element.js';
 import { TextElement } from './text-element.js';
 
 /** @typedef {import('../../../core/math/color.js').Color} Color */
-/** @typedef {import('../../../font/canvas-font.js').CanvasFont} CanvasFont */
-/** @typedef {import('../../../font/font.js').Font} Font */
+/** @typedef {import('../../font/canvas-font.js').CanvasFont} CanvasFont */
+/** @typedef {import('../../font/font.js').Font} Font */
 /** @typedef {import('../../../platform/graphics/texture.js').Texture} Texture */
 /** @typedef {import('../../../scene/materials/material.js').Material} Material */
 /** @typedef {import('../../../scene/sprite.js').Sprite} Sprite */
